+++ src/generators/model/Generator.php
@@ -12,6 +12,7 @@ use yii\base\InvalidConfigException;
 use yii\base\NotSupportedException;
 use yii\db\ActiveQuery;
 use yii\db\ActiveRecord;
+use yii\db\ColumnSchema;
 use yii\db\Connection;
 use yii\db\Exception;
 use yii\db\Schema;
@@ -307,6 +308,7 @@ class Generator extends \yii\gii\Generator
                 'rules' => $this->generateRules($tableSchema),
                 'relations' => $tableRelations,
                 'relationsClassHints' => $this->generateRelationsClassHints($tableRelations, $this->generateQuery),
+                'enum' => $this->getEnum($tableSchema->columns),
             ];
             $files[] = new CodeFile(
                 Yii::getAlias('@' . str_replace('\\', '/', $this->ns)) . '/' . $modelClassName . '.php',
@@ -441,12 +443,21 @@ class Generator extends \yii\gii\Generator
     {
         $types = [];
         $lengths = [];
+        $nullable = [];
+        $defaultValues = [];
         foreach ($table->columns as $column) {
             if ($column->autoIncrement) {
                 continue;
             }
             if (!$column->allowNull && $column->defaultValue === null) {
                 $types['required'][] = $column->name;
+            } elseif ($column->allowNull && $column->defaultValue === null) {
+                $nullable[] = $column->name;
+            } elseif (is_scalar($column->defaultValue)) {
+                if (array_key_exists($column->defaultValue, $defaultValues)) {
+                    $defaultValues[$column->defaultValue] = [];
+                }
+                $defaultValues[$column->defaultValue][] = $column->name;
             }
             switch ($column->type) {
                 case Schema::TYPE_SMALLINT:
@@ -480,6 +491,15 @@ class Generator extends \yii\gii\Generator
             }
         }
         $rules = [];
+        if (!empty($nullable)) {
+            $rules[] = "[['" . implode("', '", $nullable) . "'], 'default', 'value' => null]";
+        }
+        if (!empty($defaultValues)) {
+            foreach ($defaultValues as $defaultValue => $defaultValueColumns) {
+                $defaultValue = is_numeric($defaultValue) ? $defaultValue : "'$defaultValue'";
+                $rules[] = "[['" . implode("', '", $defaultValueColumns) . "'], 'default', 'value' => $defaultValue]";
+            }
+        }
         $driverName = $this->getDbDriverName();
         foreach ($types as $type => $columns) {
             if ($driverName === 'pgsql' && $type === 'integer') {
@@ -489,6 +509,11 @@ class Generator extends \yii\gii\Generator
         }
         foreach ($lengths as $length => $columns) {
             $rules[] = "[['" . implode("', '", $columns) . "'], 'string', 'max' => $length]";
+        }
+
+        $columnsEnum = $this->getEnum($table->columns);
+        foreach ($columnsEnum as $fieldName => $columnEnum) {
+            $rules['enum-' . $fieldName] = "['" . $fieldName . "', 'in', 'range' => array_keys(self::" . $columnEnum['funcOptsName'] . '())]';
         }
 
         $db = $this->getDbConnection();
@@ -1168,6 +1193,57 @@ class Generator extends \yii\gii\Generator
         }
 
         return false;
+    }
+
+    /**
+     * Prepares ENUM field values.
+     *
+     * @param ColumnSchema[] $columns
+     *
+     * @return array
+     */
+    public function getEnum($columns)
+    {
+        $enum = [];
+        foreach ($columns as $column) {
+            if (!$this->isEnum($column)) {
+                continue;
+            }
+
+            $columnCamelName = Inflector::id2camel($column->name, '_');
+            $enum[$column->name]['funcOptsName'] = 'opts' . $columnCamelName;
+            $enum[$column->name]['isFunctionPrefix'] = 'is' . $columnCamelName;
+            $enum[$column->name]['setFunctionPrefix'] = 'set' . $columnCamelName . 'To';
+            $enum[$column->name]['displayFunctionPrefix'] = 'display' . $columnCamelName;
+            $enum[$column->name]['columnName'] = $column->name;
+            $enum[$column->name]['values'] = [];
+
+            foreach ($column->enumValues as $value) {
+
+                $constantName = strtoupper(Inflector::slug($column->name . ' ' . $value, '_'));
+                $label = Inflector::camel2words($value);
+
+                $enum[$column->name]['values'][] = [
+                    'value' => $value,
+                    'constName' => $constantName,
+                    'label' => $label,
+                    'functionSuffix' => Inflector::id2camel(Inflector::slug($value))
+                ];
+            }
+        }
+
+        return $enum;
+    }
+
+    /**
+     * Checks if column is of ENUM type.
+     *
+     * @param ColumnSchema $column Column instance
+     * @return bool
+     */
+    protected function isEnum($column)
+    {
+        return !empty($column->enumValues) || stripos($column->dbType, 'ENUM') === 0;
     }
 
     /**